Add basic player movement and platform collision detection
Currently, the platforms and some of the game logic features are hard coded into the human view as this is only for demo purposes.

+++ include/moveable.h
@@ -5,7 +5,7 @@
 class Moveable : public Actor {
     
     public:
-        virtual void move(int distanceX, int distanceY) = 0;
+        virtual void move(float distanceX, float distanceY) = 0;
         
 
 };
